Change minimum API version to 16
I realized that not everybody's test devices support version 21 of the API
and higher. So to provide support for as many devices as possible, I am lowering
the API version from 21 to 16.
Version 16 was chosen because it was the last version that introduced a
significant change to the camera API, a very key component of this app.