Improve per-frame callback performance
Copying a preview frame into a bitmap every frame is costly. It visibly
degrades the preview's framerate. With some small revisions to the camera
capability interface, this cost can be reduced significantly. The first
is the introduction of the onPreFrame() callback that returns a Bitmap
object or null. And when null is returned, no copy is performed. The
onFrame() callback is provided the same Bitmap with the current preview
frame copied into it.