Compensate for differences in Camera.setDisplayOrientation(int) behavior
It seems that the behavior of Camera.setDisplayOrientation(int) is different
for different devices. Setting the same orientation angle on different devices
sometimes results in the image being upside down. This change seeks to fix
this issue by gathering orientation information from the camera and the
display and using that information to calculate the proper orientation angle.