Add error handling to CameraInstance
Exceptions thrown during the camera's initialization cannot be caught due to
the fact that initialization occurs on a separate thread. Here I am adding a
mechanism to handle exceptions thrown at initialization by adding a new
callback function, onCameraError(Exception). If the camera fails to initialize
and throws an exception, this function will be called instead of
onCameraStart().