Clear LayerView's screen on SurfaceTexture initialization
I ran into a strange issue in which the app's screen appeared black when
coming back from minimization. It turns out the issue had to do with
the LayerView's internal SurfaceTexture reinitializing when returning
to the app. The color of the new SurfaceTexture was completely black.
Filling the SurfaceTexture with transparent pixels on every initialization
solves this issue.

Also, I removed LayerView.init() and replaced it with an initialization block.
I would have separated this change into its own commit, but it proved to be
too difficult to separate it from the other changes I made in the same file.