Remove 'static' calls to Cursor.getColumnIndex()
It occurred to me that the column index for a field can be different
for different queries. Consider the following SQL queries
    SELECT key, name FROM identities

    SELECT name, key FROM identities
The column indexes for the 'key' and 'name' fields are different
between the top and bottom queries. I had misunderstood that
Cursor.getColumnIndex() returned the column index with respect
to the table's schema. And considering that the table schema does
not change (aside from schema updates), I thought I could improve
performance by calling this method only once for each column.